More conditions on shutil.rmtree, fixed params for validation repo urls

+++ tests/test_arca_class.py
@@ -1,4 +1,6 @@
 # encoding=utf-8
+import os
+import platform
 import re
 import shutil
 from pathlib import Path
@@ -32,18 +34,33 @@ def test_arca_backend():
 
 
 @pytest.mark.parametrize(["url", "valid"], [
+    # http/s
     ("http://host.xz/path/to/repo.git/", True),
     ("https://host.xz/path/to/repo.git/", True),
     ("http://host.xz/path/to/repo.git", True),
     ("https://host.xz/path/to/repo.git", True),
-    ("file:///path/to/repo.git/", True),
-    ("file://~/path/to/repo.git/", True),
     ("http://host.xz/path/to/repo/", True),
     ("https://host.xz/path/to/repo/", True),
     ("http://host.xz/path/to/repo", True),
     ("https://host.xz/path/to/repo", True),
-    ("file:///path/to/repo.git", True),
-    ("file://~/path/to/repo.git", True),
+
+    # linux paths
+    pytest.param("file:///path/to/repo.git/", True,
+                 marks=pytest.mark.skipif(platform.system() == "Windows", reason="Linux Path")),
+    pytest.param("file://~/path/to/repo.git/", True,
+                 marks=pytest.mark.skipif(platform.system() == "Windows", reason="Linux Path")),
+    pytest.param("file:///path/to/repo.git", True,
+                 marks=pytest.mark.skipif(platform.system() == "Windows", reason="Linux Path")),
+    pytest.param("file://~/path/to/repo.git", True,
+                 marks=pytest.mark.skipif(platform.system() == "Windows", reason="Linux Path")),
+
+    # windows paths
+    pytest.param("file:///C:\\user\\path \\to\\repo", True,
+                 marks=pytest.mark.skipif(platform.system() != "Windows", reason="Windows Path")),
+    pytest.param("http:///c:\\user\\path \\to\\repo", True,
+                 marks=pytest.mark.skipif(platform.system() != "Windows", reason="Windows Path")),
+
+    # ssh
     ("git://host.xz/path/to/repo.git/", False),
     ("git://host.xz/~user/path/to/repo.git/", False),
     ("ssh://host.xz/path/to/repo.git/", False),
@@ -177,7 +194,8 @@ def test_depth(temp_repo_static):
     cloned_repo, cloned_repo_path = arca.get_files(temp_repo_static.url, temp_repo_static.branch)
     assert cloned_repo.commit().count() == 2
 
-    shutil.rmtree(str(cloned_repo_path))
+    if not os.environ.get("APPVEYOR", False):
+        shutil.rmtree(str(cloned_repo_path))
 
     # test that when setting a certain depth, at least the depth is pulled (in case of merges etc)
 
@@ -194,7 +212,8 @@ def test_depth(temp_repo_static):
     cloned_repo, cloned_repo_path = arca.get_files(temp_repo_static.url, temp_repo_static.branch)
     assert cloned_repo.commit().count() == before_second_pull + 1
 
-    shutil.rmtree(str(cloned_repo_path))
+    if not os.environ.get("APPVEYOR", False):
+        shutil.rmtree(str(cloned_repo_path))
 
     # test when setting depth bigger than repo size, no fictional commits are included
 
@@ -202,7 +221,8 @@ def test_depth(temp_repo_static):
 
     assert cloned_repo.commit().count() == 22  # 20 plus the 2 extra commits
 
-    shutil.rmtree(str(cloned_repo_path))
+    if not os.environ.get("APPVEYOR", False):
+        shutil.rmtree(str(cloned_repo_path))
 
     # test no limit
 
@@ -254,7 +274,9 @@ def test_reference():
 
     cloned_repo, cloned_repo_path = arca.get_files(git_url_1, branch, reference=Path("/tmp/arca/") / str(uuid4()))
     assert (cloned_repo_path / "test_file.txt").read_text() == last_uuid
-    shutil.rmtree(str(cloned_repo_path))
+
+    if not os.environ.get("APPVEYOR", False):
+        shutil.rmtree(str(cloned_repo_path))
 
     # test existing reference with no common commits
 
@@ -269,7 +291,9 @@ def test_reference():
 
     cloned_repo, cloned_repo_path = arca.get_files(git_url_1, branch, reference=git_dir_2)
     assert (cloned_repo_path / "test_file.txt").read_text() == last_uuid
-    shutil.rmtree(str(cloned_repo_path))
+
+    if not os.environ.get("APPVEYOR", False):
+        shutil.rmtree(str(cloned_repo_path))
 
     # test existing reference with common commits
 
@@ -351,7 +375,8 @@ def test_pull_error():
     with pytest.raises(PullError):
         arca.get_files(git_url, "some_branch")
 
-    shutil.rmtree(str(git_dir))
+    if not os.environ.get("APPVEYOR", False):
+        shutil.rmtree(str(git_dir))
 
     with pytest.raises(PullError):
         arca.get_files(git_url, "master")
